+++ forge/src/main/java/club/iananderson/seasonhud/impl/minimaps/CurrentMinimap.java
@@ -4,11 +4,12 @@ import club.iananderson.seasonhud.platform.Services;
 import net.minecraft.client.Minecraft;
 import net.minecraft.resources.ResourceKey;
 import net.minecraft.world.level.Level;
+import sereneseasons.init.ModConfig;
+
 import java.util.Objects;
 
 import static club.iananderson.seasonhud.config.Config.enableMod;
 import static club.iananderson.seasonhud.impl.seasons.Calendar.calendar;
-import static net.minecraft.world.level.Level.OVERWORLD;
 
 public class CurrentMinimap {
     public static boolean loadedMinimap(String minimap){
@@ -32,6 +33,6 @@ public class CurrentMinimap {
         Minecraft mc = Minecraft.getInstance();
         ResourceKey<Level> currentDim = Objects.requireNonNull(mc.level).dimension();
 
-        return currentDim != OVERWORLD;
+        return ModConfig.seasons.isDimensionWhitelisted(currentDim);
     }
 }