+++ neoforge/src/main/java/club/iananderson/seasonhud/client/SeasonHUDOverlay.java
@@ -1,0 +1,82 @@
+package club.iananderson.seasonhud.client;
+
+import club.iananderson.seasonhud.config.Location;
+import net.minecraft.client.Minecraft;
+import net.minecraft.client.gui.Font;
+import net.minecraft.client.gui.GuiGraphics;
+import net.minecraft.client.gui.screens.ChatScreen;
+import net.minecraft.client.gui.screens.DeathScreen;
+import net.minecraft.network.chat.Component;
+import net.minecraft.network.chat.MutableComponent;
+import net.neoforged.neoforge.client.gui.overlay.ExtendedGui;
+import net.neoforged.neoforge.client.gui.overlay.IGuiOverlay;
+
+import static club.iananderson.seasonhud.Common.SEASON_STYLE;
+import static club.iananderson.seasonhud.config.Config.*;
+import static club.iananderson.seasonhud.impl.minimaps.CurrentMinimap.noMinimap;
+import static club.iananderson.seasonhud.impl.minimaps.HiddenMinimap.minimapHidden;
+import static club.iananderson.seasonhud.impl.seasons.Calendar.calendar;
+import static club.iananderson.seasonhud.impl.seasons.CurrentSeason.getSeasonName;
+
+public class SeasonHUDOverlay {
+
+    public static void render(GuiGraphics seasonStack, float partialTick) {
+        Minecraft mc = Minecraft.getInstance();
+
+        MutableComponent seasonCombined = Component.translatable("desc.seasonhud.combined",
+                getSeasonName().get(0).copy().withStyle(SEASON_STYLE),
+                getSeasonName().get(1).copy());
+
+        float guiSize = (float) mc.getWindow().getGuiScale();
+
+        int screenWidth = mc.getWindow().getGuiScaledWidth();
+        int screenHeight = mc.getWindow().getGuiScaledHeight();
+
+        int xOffset = (int) (hudX.get()/guiSize);
+        int yOffset = (int) ((hudY.get())/guiSize);
+        int x = 1;
+        int y = 1;
+        int offsetDim = 2;
+
+        Font font = mc.font;
+        int stringWidth = font.width(seasonCombined);
+
+        if ((noMinimap() || (minimapHidden() && showMinimapHidden.get())) && enableMod.get() && calendar()) {
+            Location hudLoc = hudLocation.get();
+            switch (hudLoc) {
+                case TOP_LEFT -> {
+                    x = offsetDim;
+                    y = 0;
+                }
+                case TOP_CENTER -> {
+                    x = screenWidth / 2 - stringWidth / 2;
+                    y = 0;
+                }
+                case TOP_RIGHT -> {
+                    x = screenWidth - stringWidth - offsetDim;
+                    y = 0;
+                }
+                case BOTTOM_LEFT -> {
+                    x = offsetDim;
+                    y = screenHeight - (2*offsetDim);
+                }
+                case BOTTOM_RIGHT -> {
+                    x = screenWidth - stringWidth - offsetDim;
+                    y = screenHeight - (2*offsetDim);
+                }
+            }
+
+            if ((mc.screen == null || mc.screen instanceof ChatScreen || mc.screen instanceof DeathScreen) && !mc.options.hideGui) {
+                seasonStack.pose().pushPose();
+                seasonStack.pose().scale(1F, 1F, 1F);
+
+                //Text
+                int iconX = x + xOffset;
+                int iconY = y + yOffset + offsetDim;
+
+                seasonStack.drawString(font, seasonCombined, iconX, iconY, 0xffffffff);
+                seasonStack.pose().popPose();
+            }
+        }
+    };
+}