+++ neoforge/src/main/java/club/iananderson/seasonhud/event/ClientEvents.java
@@ -1,0 +1,47 @@
+package club.iananderson.seasonhud.event;
+
+import club.iananderson.seasonhud.SeasonHUD;
+import club.iananderson.seasonhud.client.KeyBindings;
+import club.iananderson.seasonhud.client.SeasonHUDOverlay;
+import club.iananderson.seasonhud.client.minimaps.JourneyMap;
+import club.iananderson.seasonhud.client.minimaps.MapAtlases;
+import net.minecraft.resources.ResourceLocation;
+import net.neoforged.api.distmarker.Dist;
+import net.neoforged.bus.api.EventPriority;
+import net.neoforged.bus.api.SubscribeEvent;
+import net.neoforged.fml.common.Mod;
+import net.neoforged.fml.common.Mod.EventBusSubscriber;
+import net.neoforged.neoforge.client.event.InputEvent;
+import net.neoforged.neoforge.client.event.RegisterGuiOverlaysEvent;
+import net.neoforged.neoforge.client.event.RegisterKeyMappingsEvent;
+import net.neoforged.neoforge.client.event.RenderGuiOverlayEvent;
+import net.neoforged.neoforge.client.gui.overlay.VanillaGuiOverlay;
+
+public class ClientEvents{
+    @Mod.EventBusSubscriber(modid = SeasonHUD.MODID, value = Dist.CLIENT)
+    public static class ClientForgeEvents {
+        @SubscribeEvent
+        public static void onKeyInput(InputEvent.Key Event) {
+            if (KeyBindings.seasonhudOptionsKeyMapping.consumeClick()) {
+                SeasonHUDScreen.open();
+            }
+        }
+    }
+
+    @Mod.EventBusSubscriber(modid = SeasonHUD.MODID, value = Dist.CLIENT, bus = Mod.EventBusSubscriber.Bus.MOD)
+    public static class ClientModBusEvents {
+        //Overlays
+        @SubscribeEvent
+        public static void renderOverlay(RenderGuiOverlayEvent.Post event) {
+            SeasonHUDOverlay.render(event.getGuiGraphics(),event.getPartialTick());
+            JourneyMap.render(event.getGuiGraphics(),event.getPartialTick());
+            MapAtlases.render(event.getGuiGraphics(),event.getPartialTick());
+        }
+
+        @SubscribeEvent
+        //Key Bindings
+        public static void onKeyRegister(RegisterKeyMappingsEvent event){
+            event.register(KeyBindings.seasonhudOptionsKeyMapping);
+        }
+    }
+}